fixed minor issue & update some style

+++ src/Pages/MyOrder/MyOrder.jsx
@@ -1,16 +1,16 @@
 import { useContext, useEffect, useState } from "react";
 import { AuthContext } from "../../Provider/AuthProvider";
-import OrderRow from "./OrderRow";
 import axios from "axios";
 import { Helmet } from "react-helmet";
+import OrderCard from "./OrderCard";
+import Swal from "sweetalert2";
 
 const MyOrder = () => {
   const { user } = useContext(AuthContext);
 
   const [orders, setOrders] = useState([]);
-  console.log(orders);
 
-  const url = `http://localhost:5000/usersFood?email=${user?.email}`;
+  const url = `http://localhost:5000/orderFoods?email=${user?.email}`;
 
   useEffect(() => {
     axios.get(url).then((res) => {
@@ -19,34 +19,35 @@ const MyOrder = () => {
   }, [url]);
 
   const deleteHandler = (id) => {
-    fetch(`http://localhost:5000/orders/${id}`, {
-      method: "DELETE",
-    })
-      .then((res) => res.json())
-      .then((data) => {
-        console.log(data);
-        const remaining = orders.filter((order) => order._id !== id);
-        setOrders(remaining);
-      });
-  };
+    Swal.fire({
+      title: "Are you sure?",
+      text: "You won't be able to revert this!",
+      icon: "warning",
+      showCancelButton: true,
+      confirmButtonColor: "#3085d6",
+      cancelButtonColor: "#d33",
+      confirmButtonText: "Yes, delete it!",
+    }).then((result) => {
+      if (result.isConfirmed) {
+        fetch(`http://localhost:5000/orderFoods/${id}`, {
+          method: "DELETE",
+        })
+          .then((res) => res.json())
+          .then((data) => {
+            console.log(data);
 
-  const handleOrderConfirm = (id) => {
-    fetch(`http://localhost:5000/orders/${id}`, {
-      method: "PATCH",
-      headers: {
-        "content-type": "application/json",
-      },
-      body: JSON.stringify({ status: "confirm" }),
-    })
-      .then((res) => res.json())
-      .then((data) => {
-        console.log(data);
-        const remaining = orders.filter((order) => order._id !== id);
-        const updated = orders.find((order) => order._id === id);
-        updated.status = "confirm";
-        const updatedOrder = [updated, ...remaining];
-        setOrders(updatedOrder);
-      });
+            if (data.deletedCount > 0) {
+              Swal.fire({
+                title: "Deleted!",
+                text: "Your file has been deleted.",
+                icon: "success",
+              });
+            }
+            const remaining = orders.filter((order) => order._id !== id);
+            setOrders(remaining);
+          });
+      }
+    });
   };
 
   return (
@@ -54,34 +55,18 @@ const MyOrder = () => {
       <Helmet>
         <title> DishDynamo | MyOrder </title>
       </Helmet>
-      <div className='overflow-x-auto'>
-        <table className='table'>
-          {/* head */}
-          <thead>
-            <tr>
-              <th>
-                <label>
-                  <input type='checkbox' className='checkbox' />
-                </label>
-              </th>
-              <th>Name</th>
-              <th>Email</th>
-              <th>Price</th>
-              <th>Action</th>
-            </tr>
-          </thead>
-          <tbody>
-            {/* row 1 */}
-            {orders.map((order) => (
-              <OrderRow
-                key={order._id}
-                order={order}
-                deleteHandler={deleteHandler}
-                handleOrderConfirm={handleOrderConfirm}
-              ></OrderRow>
-            ))}
-          </tbody>
-        </table>
+      <div className='grid md:grid-cols-2 gap-5 container mx-auto py-10'>
+        {orders.length === 0 ? (
+          <p className='col-span-2 text-center'>No Order Product</p>
+        ) : (
+          orders.map((order) => (
+            <OrderCard
+              deleteHandler={deleteHandler}
+              key={order._id}
+              order={order}
+            ></OrderCard>
+          ))
+        )}
       </div>
     </div>
   );
